--- asjp/tests/conftest.py
@@ -1,3 +1,0 @@
-import pytest
-
-pytest_plugins = ['clld']
