+++ development.ini
@@ -1,0 +1,43 @@
+[app:main]
+use = egg:asjp
+
+pyramid.reload_templates = true
+pyramid.debug_authorization = false
+pyramid.debug_notfound = false
+pyramid.debug_routematch = false
+pyramid.default_locale_name = en
+pyramid.includes =
+    pyramid_tm
+sqlalchemy.url = postgresql://postgres@/asjp
+
+[server:main]
+use = egg:waitress#main
+host = 127.0.0.1
+port = 6543
+
+[loggers]
+keys = root, asjp
+
+[handlers]
+keys = console
+
+[formatters]
+keys = generic
+
+[logger_root]
+level = INFO
+handlers = console
+
+[logger_asjp]
+level = DEBUG
+handlers =
+qualname = asjp
+
+[handler_console]
+class = StreamHandler
+args = (sys.stderr,)
+level = NOTSET
+formatter = generic
+
+[formatter_generic]
+format = %(asctime)s %(levelname)-5.5s [%(name)s][%(threadName)s] %(message)s
