--- fabfile.py
@@ -1,2 +1,0 @@
-from clldfabric import tasks
-tasks.init('asjp')
