+++ setup.cfg
@@ -6,6 +6,12 @@ addopts =
     --cov=asjp
     --cov-report term-missing
 
+[coverage:run]
+source = asjp
+
+[coverage:report]
+show_missing = true
+
 [compile_catalog]
 directory = asjp/locale
 domain = clld
